refactor: move font select to Shared folder
